add Collection Controller and update collection model.
